Update settlement event layout
